fix(action): honour an extra-args --output override when locating the report
`extra-args` can carry its own `-o`/`--output`, and since `CMD` puts this
script's own flags first and `$INPUT_EXTRA_ARGS` last, Click's last-wins rule
makes the override the real destination -- the same override
`_effective_format` already existed to resolve for `--format`. Keying the
destination inventory on `$OUTPUT_FILE` alone was wrong in both directions
(Codex review, P2, reproduced):

* with no `output-file` input, `format: json` plus `extra-args: --output
  report.json` named no destination, so the run was validated as the stdout
  shape -- where the capture is empty, because output went to a file -- and a
  working run published REPORT_UNREADABLE, a regression this PR introduced;
* with both given, the inventory validated the superseded path while the real
  report landed elsewhere.

`_effective_output_file` resolves it, and `_json_report_src` and the pre-run
fingerprint bookkeeping resolve the same path. Fixing the reading chain is
not optional here: had only the inventory learned the effective path, the
report would validate and then never be read, falling through to COMPATIBLE
-- this PR's original defect. Mutation-tested in exactly that split: reverting
both kills 7 tests, reverting only the chain kills 4.

Also splits `test_action_unreadable_report_verdict.py`, which crossed the
1200-line test cap. Split by responsibility rather than trimmed to fit, per
the root AGENTS.md: the destination questions (where a report was asked for,
and whether this run wrote it) move to `test_action_report_destinations.py`,
and the shared stub-and-runner harness both need moves to
`tests/_action_run_sh_harness.py` rather than being duplicated. The harness
owns `REQUIRES_POSIX_SHELL` too -- one sibling module having that marker and
the other not is what turned a windows-latest lane red earlier on this PR.
