+++ tests/test_action_unreadable_report_verdict.py
@@ -39,115 +39,19 @@ Bug class: ``report.unestablished_result_reads_as_success``
 from __future__ import annotations
 
 import json
-import os
-import shutil
-import subprocess
 from pathlib import Path
 
 import pytest
-
-ACTION_DIR = Path(__file__).resolve().parents[1] / "action"
-RUN_SH = ACTION_DIR / "run.sh"
-
-# POSIX only, for the same reason `test_action_coverage_verdict.py` states: this
-# harness works by putting an *extensionless, shebang-dispatched* `abicheck` on
-# PATH, because `run.sh` resolves the binary by name. Windows has neither the
-# executable bit nor kernel shebang handling, and Git bash's `chmod` is a no-op
-# on NTFS, so the stub is not runnable there — every test in the module then
-# fails identically with the WSL launcher stub's own UTF-16 "no installed
-# distributions" text instead of anything from `run.sh`. The behaviour under
-# test is plain shell with no platform-dependent branch, and the Linux lane
-# exercises all of it.
-#
-# Omitting this marker is what turned the windows-latest unit lane red on this
-# PR: the sibling module documented the pitfall and this one did not copy it.
-pytestmark = pytest.mark.skipif(
-    os.name == "nt" or not RUN_SH.is_file() or shutil.which("bash") is None,
-    reason="needs a POSIX shell that can exec a shebang script from PATH",
+from _action_run_sh_harness import (
+    REQUIRES_POSIX_SHELL,
+    _compare_env,
+    _lib,
+    _run_action,
+    _stub_abicheck,
+    _stub_stdout_only,
 )
 
-
-def _stub_abicheck(tmp_path: Path, *, exit_code: int, payload: bytes | None) -> Path:
-    """An abicheck that exits *exit_code* and writes *payload* to its ``-o`` path.
-
-    ``payload=None`` writes nothing at all -- the "died after the exit code,
-    before the report" shape, which is the one a real truncated/killed run
-    produces and which no valid-JSON fixture can stand in for.
-    """
-    bindir = tmp_path / "bin"
-    bindir.mkdir(exist_ok=True)
-    body = [
-        "#!/usr/bin/env bash",
-        "prev=''",
-        'for arg in "$@"; do',
-        '  if [[ "$prev" == "-o" ]]; then',
-    ]
-    if payload is None:
-        body.append("    :")
-    else:
-        blob = tmp_path / "payload.bin"
-        blob.write_bytes(payload)
-        body.append(f'    cp "{blob}" "$arg"')
-    body += ["  fi", '  prev="$arg"', "done", f"exit {exit_code}"]
-    stub = bindir / "abicheck"
-    stub.write_text("\n".join(body) + "\n", encoding="utf-8")
-    stub.chmod(0o755)
-    return bindir
-
-
-def _lib(tmp_path: Path, name: str) -> str:
-    path = tmp_path / name
-    path.write_bytes(b"\x7fELF")
-    return str(path)
-
-
-def _run_action(tmp_path: Path, env_extra: dict[str, str], bindir: Path) -> dict:
-    out = tmp_path / "github_output"
-    out.write_text("", encoding="utf-8")
-    summary = tmp_path / "step_summary"
-    summary.write_text("", encoding="utf-8")
-    runner_temp = tmp_path / "runner_temp"
-    runner_temp.mkdir(exist_ok=True)
-    env = {k: v for k, v in os.environ.items() if not k.startswith("INPUT_")}
-    env.update(
-        {
-            "PATH": f"{bindir}{os.pathsep}{env.get('PATH', '')}",
-            "ACTION_PATH": str(ACTION_DIR),
-            "GITHUB_OUTPUT": str(out),
-            "GITHUB_STEP_SUMMARY": str(summary),
-            "RUNNER_TEMP": str(runner_temp),
-            "INPUT_ADD_JOB_SUMMARY": "true",
-            **env_extra,
-        }
-    )
-    proc = subprocess.run(
-        ["bash", str(RUN_SH)],
-        capture_output=True,
-        text=True,
-        env=env,
-        cwd=tmp_path,
-        check=False,
-    )
-    outputs: dict = {}
-    for line in out.read_text(encoding="utf-8").splitlines():
-        if "=" in line:
-            key, value = line.split("=", 1)
-            outputs[key] = value
-    outputs["_stdout"] = proc.stdout
-    outputs["_exit"] = proc.returncode
-    outputs["_summary"] = summary.read_text(encoding="utf-8")
-    return outputs
-
-
-def _compare_env(tmp_path: Path) -> dict[str, str]:
-    return {
-        "INPUT_MODE": "compare",
-        "INPUT_OLD_LIBRARY": _lib(tmp_path, "libold.so"),
-        "INPUT_NEW_LIBRARY": _lib(tmp_path, "libnew.so"),
-        "INPUT_FORMAT": "json",
-        "INPUT_OUTPUT_FILE": str(tmp_path / "report.json"),
-    }
-
+pytestmark = REQUIRES_POSIX_SHELL
 
 #: Each is a distinct way for an exit-0 run to leave no readable result. The
 #: reported instance was ``{}``; enumerating the class is the point -- a fix
@@ -458,27 +362,6 @@ class TestTheScopeOfTheCheckIsDeliberate:
         assert outputs["_exit"] == 1, outputs
 
 
-def _stub_stdout_only(tmp_path: Path, *, stdout: str) -> Path:
-    """An abicheck that writes nothing to ``-o`` and prints *stdout* instead.
-
-    The documented `format: json` stdout mode -- no `output-file` at all, the
-    report goes to stdout. An empty *stdout* is the failure this shape can
-    reach.
-    """
-    bindir = tmp_path / "bin"
-    bindir.mkdir(exist_ok=True)
-    stub = bindir / "abicheck"
-    body = "#!/usr/bin/env bash\n"
-    if stdout:
-        blob = tmp_path / "stdout.txt"
-        blob.write_text(stdout, encoding="utf-8")
-        body += f'cat "{blob}"\n'
-    body += "exit 0\n"
-    stub.write_text(body, encoding="utf-8")
-    stub.chmod(0o755)
-    return bindir
-
-
 class TestEveryCallerRequestedJsonModeIsCovered:
     """`format: json` is the request; where the report lands is a separate choice.
 
@@ -546,130 +429,6 @@ class TestEveryCallerRequestedJsonModeIsCovered:
                 INPUT_EXTRA_ARGS=f"--write json={target}",
             ),
             bindir,
-        )
-        assert outputs["verdict"] == "REPORT_UNREADABLE", outputs
-        assert outputs["_exit"] == 1, outputs
-
-
-class TestEveryRepeatableWriteDestinationIsChecked:
-    """`--write` is repeatable, so every `json=` destination is a real request.
-
-    `compare` declares `--write` with `multiple=True` (ADR-068 D4, "one
-    analysis, several artifacts"), and every named artifact is written. The
-    extractor behind this predicate used to *clear* an already-found `json=`
-    path whenever a later `--write` named another format — matching a stale
-    comment that called the option scalar and last-wins. So
-    `--write json=a.json --write markdown=b.md` reported no requested JSON path
-    at all, and a missing `a.json` left an exit-0 run publishing COMPATIBLE
-    (Codex review, P2).
-
-    The two contradictory claims in the tree were settled against the option
-    declaration itself, not either comment.
-    """
-
-    def _env(self, tmp_path: Path, extra_args: str) -> dict[str, str]:
-        return {
-            "INPUT_MODE": "compare",
-            "INPUT_OLD_LIBRARY": _lib(tmp_path, "libold.so"),
-            "INPUT_NEW_LIBRARY": _lib(tmp_path, "libnew.so"),
-            "INPUT_FORMAT": "markdown",
-            "INPUT_EXTRA_ARGS": extra_args,
-        }
-
-    def _stub_writing(self, tmp_path: Path, *, honor: set[str]) -> Path:
-        """An abicheck that writes only the `json=` destinations in *honor*.
-
-        Lets a test name two JSON destinations and have exactly one arrive —
-        the shape the clearing bug hid.
-        """
-        bindir = tmp_path / "bin"
-        bindir.mkdir(exist_ok=True)
-        payload = json.dumps({"report_schema_version": "4.4", "verdict": "COMPATIBLE"})
-        lines = ["#!/usr/bin/env bash"]
-        for name in sorted(honor):
-            lines.append(f"printf '%s' '{payload}' > {name}")
-        lines.append("exit 0")
-        stub = bindir / "abicheck"
-        stub.write_text("\n".join(lines) + "\n", encoding="utf-8")
-        stub.chmod(0o755)
-        return bindir
-
-    def test_a_json_write_followed_by_another_format_is_still_required(
-        self, tmp_path: Path
-    ) -> None:
-        # The exact reported case: the json destination is named first, a
-        # non-json `--write` follows, and nothing writes the json one.
-        target = tmp_path / "a.json"
-        bindir = self._stub_writing(tmp_path, honor=set())
-        outputs = _run_action(
-            tmp_path,
-            self._env(
-                tmp_path, f"--write json={target} --write markdown={tmp_path / 'b.md'}"
-            ),
-            bindir,
-        )
-        assert outputs["verdict"] == "REPORT_UNREADABLE", outputs
-        assert outputs["_exit"] == 1, outputs
-
-    def test_the_same_combination_passes_when_the_json_destination_arrives(
-        self, tmp_path: Path
-    ) -> None:
-        # Negative control: identical extra-args, but the stub honors the json
-        # destination. Without this, the fix is satisfiable by rejecting every
-        # multi-write invocation.
-        target = tmp_path / "a.json"
-        bindir = self._stub_writing(tmp_path, honor={str(target)})
-        outputs = _run_action(
-            tmp_path,
-            self._env(
-                tmp_path, f"--write json={target} --write markdown={tmp_path / 'b.md'}"
-            ),
-            bindir,
-        )
-        assert outputs["verdict"] == "COMPATIBLE", outputs
-        assert outputs["_exit"] == 0, outputs
-
-    def test_a_second_json_destination_is_checked_too(self, tmp_path: Path) -> None:
-        # "Track all caller-supplied JSON destinations" — one arriving does not
-        # excuse another that did not, so a readable first destination must not
-        # mask an absent second.
-        first = tmp_path / "a.json"
-        second = tmp_path / "b.json"
-        bindir = self._stub_writing(tmp_path, honor={str(first)})
-        outputs = _run_action(
-            tmp_path,
-            self._env(tmp_path, f"--write json={first} --write json={second}"),
-            bindir,
-        )
-        assert outputs["verdict"] == "REPORT_UNREADABLE", outputs
-        assert outputs["_exit"] == 1, outputs
-
-    def test_both_json_destinations_arriving_passes(self, tmp_path: Path) -> None:
-        first = tmp_path / "a.json"
-        second = tmp_path / "b.json"
-        bindir = self._stub_writing(tmp_path, honor={str(first), str(second)})
-        outputs = _run_action(
-            tmp_path,
-            self._env(tmp_path, f"--write json={first} --write json={second}"),
-            bindir,
-        )
-        assert outputs["verdict"] == "COMPATIBLE", outputs
-        assert outputs["_exit"] == 0, outputs
-
-    def test_a_resultless_destination_is_caught(self, tmp_path: Path) -> None:
-        target = tmp_path / "a.json"
-        bindir = tmp_path / "bin"
-        bindir.mkdir(exist_ok=True)
-        stub = bindir / "abicheck"
-        stub.write_text(
-            "#!/usr/bin/env bash\n"
-            f"printf '%s' '{{\"error\": \"write interrupted\"}}' > {target}\n"
-            "exit 0\n",
-            encoding="utf-8",
-        )
-        stub.chmod(0o755)
-        outputs = _run_action(
-            tmp_path, self._env(tmp_path, f"--write json={target}"), bindir
         )
         assert outputs["verdict"] == "REPORT_UNREADABLE", outputs
         assert outputs["_exit"] == 1, outputs
@@ -791,323 +550,3 @@ class TestADryRunIsNotAMissingReport:
         outputs = self._dry_run(tmp_path)
         assert outputs["verdict"] == "REPORT_UNREADABLE", outputs
         assert outputs["_exit"] == 1, outputs
-
-
-#: Every way one named destination can fail to hold a usable report *this run*
-#: produced. `stale` is the odd one out and the point of the enumeration: the
-#: document is valid, parses, carries a real verdict, and was written by some
-#: earlier run — indistinguishable from a good report by content alone.
-UNUSABLE_DESTINATION_STATES = (
-    "absent",
-    "zero bytes",
-    "empty object",
-    "no result",
-    "truncated",
-    "stale",
-)
-
-_GOOD_REPORT = json.dumps({"report_schema_version": "4.4", "verdict": "COMPATIBLE"})
-
-_STATE_PAYLOADS = {
-    "zero bytes": "",
-    "empty object": "{}",
-    "no result": '{"error": "write interrupted"}',
-    "truncated": '{"verdict": "COMPATIBLE"',
-}
-
-
-class TestEveryRequestedDestinationIsJudgedOnItsOwn:
-    """A requested destination must hold a report *this invocation* produced.
-
-    Two review findings (Codex, P2, both reproduced) with one root cause: the
-    validation asked "is there a readable report somewhere" rather than "did
-    this run produce a usable report at every destination it was asked for".
-
-    * `_report_validity` followed `_json_report_src`, which is a *fallback
-      chain* — so a missing `format: json` primary was masked by a valid
-      `extra-args --write json=secondary.json`, and the step published a
-      compatibility verdict though the requested output never arrived.
-    * the per-`--write` loop checked parseability without freshness — so a
-      leftover document (or one a PR author committed, `extra-args` being
-      PR-controlled) read as one this run had written.
-
-    The invariant, stated over the cross-product rather than the two reported
-    inputs: *for any single requested destination left in any unusable state,
-    with every other destination valid and freshly written, the step must
-    publish REPORT_UNREADABLE and fail.* One destination arriving never
-    answers for another, and content alone never establishes authorship.
-    """
-
-    def _paths(self, tmp_path: Path) -> tuple[Path, Path]:
-        return tmp_path / "primary.json", tmp_path / "secondary.json"
-
-    def _stub(self, tmp_path: Path, writes: dict[Path, str]) -> Path:
-        """An abicheck writing exactly *writes* (destination -> literal text)."""
-        bindir = tmp_path / "bin"
-        bindir.mkdir(exist_ok=True)
-        lines = ["#!/usr/bin/env bash"]
-        for index, (dest, text) in enumerate(sorted(writes.items())):
-            # Via a file, so no payload text has to survive shell quoting.
-            blob = tmp_path / f"blob{index}.bin"
-            blob.write_text(text, encoding="utf-8")
-            lines.append(f'cp "{blob}" "{dest}"')
-        lines.append("exit 0")
-        stub = bindir / "abicheck"
-        stub.write_text("\n".join(lines) + "\n", encoding="utf-8")
-        stub.chmod(0o755)
-        return bindir
-
-    def _env(self, tmp_path: Path, primary: Path, secondary: Path) -> dict[str, str]:
-        return {
-            "INPUT_MODE": "compare",
-            "INPUT_OLD_LIBRARY": _lib(tmp_path, "libold.so"),
-            "INPUT_NEW_LIBRARY": _lib(tmp_path, "libnew.so"),
-            "INPUT_FORMAT": "json",
-            "INPUT_OUTPUT_FILE": str(primary),
-            "INPUT_EXTRA_ARGS": f"--write json={secondary}",
-        }
-
-    def _arrange(
-        self, tmp_path: Path, broken: str, state: str
-    ) -> tuple[Path, dict[str, str]]:
-        """Set up so *broken* ("primary"/"secondary") is in *state*, the other good."""
-        primary, secondary = self._paths(tmp_path)
-        target = primary if broken == "primary" else secondary
-        other = secondary if broken == "primary" else primary
-        writes = {other: _GOOD_REPORT}
-        if state == "stale":
-            # Pre-exists with perfectly valid content this run never wrote.
-            target.write_text(_GOOD_REPORT, encoding="utf-8")
-        elif state != "absent":
-            writes[target] = _STATE_PAYLOADS[state]
-        return self._stub(tmp_path, writes), self._env(tmp_path, primary, secondary)
-
-    @pytest.mark.parametrize("broken", ("primary", "secondary"))
-    @pytest.mark.parametrize("state", UNUSABLE_DESTINATION_STATES)
-    def test_one_unusable_destination_is_never_masked_by_the_other(
-        self, tmp_path: Path, broken: str, state: str
-    ) -> None:
-        bindir, env = self._arrange(tmp_path, broken, state)
-        outputs = _run_action(tmp_path, env, bindir)
-        assert outputs.get("verdict") == "REPORT_UNREADABLE", (broken, state, outputs)
-        assert outputs["_exit"] == 1, (broken, state, outputs)
-
-    @pytest.mark.parametrize("broken", ("primary", "secondary"))
-    @pytest.mark.parametrize("state", UNUSABLE_DESTINATION_STATES)
-    def test_the_diagnostic_names_the_destination_that_failed(
-        self, tmp_path: Path, broken: str, state: str
-    ) -> None:
-        # A verdict alone leaves the user hunting; and naming the *other*
-        # destination is exactly the confusion the fallback chain caused.
-        #
-        # Scoped to the `::error::` annotations, not to stdout as a whole: every
-        # destination path also appears in the echoed command line, so a
-        # substring search over all output passes even when no diagnostic was
-        # emitted at all (caught by mutation-testing this assertion).
-        bindir, env = self._arrange(tmp_path, broken, state)
-        outputs = _run_action(tmp_path, env, bindir)
-        primary, secondary = self._paths(tmp_path)
-        named = primary if broken == "primary" else secondary
-        other = secondary if broken == "primary" else primary
-        errors = [
-            line
-            for line in outputs["_stdout"].splitlines()
-            if line.startswith("::error::")
-        ]
-        blamed = [line for line in errors if str(named) in line]
-        assert blamed, (broken, state, errors)
-        assert not [line for line in errors if str(other) in line], (
-            broken,
-            state,
-            errors,
-        )
-
-    def test_both_destinations_freshly_written_is_compatible(
-        self, tmp_path: Path
-    ) -> None:
-        # The control that keeps every assertion above from being satisfiable
-        # by rejecting any multi-destination invocation outright.
-        primary, secondary = self._paths(tmp_path)
-        bindir = self._stub(tmp_path, {primary: _GOOD_REPORT, secondary: _GOOD_REPORT})
-        outputs = _run_action(tmp_path, self._env(tmp_path, primary, secondary), bindir)
-        assert outputs.get("verdict") == "COMPATIBLE", outputs
-        assert outputs["_exit"] == 0, outputs
-
-    def test_a_pre_existing_destination_the_run_overwrites_is_accepted(
-        self, tmp_path: Path
-    ) -> None:
-        # Freshness must key on "did this run write it", not "did it pre-exist":
-        # re-running a workflow step over an existing report.json is ordinary.
-        primary, secondary = self._paths(tmp_path)
-        primary.write_text(json.dumps({"verdict": "BREAKING"}), encoding="utf-8")
-        secondary.write_text(json.dumps({"verdict": "BREAKING"}), encoding="utf-8")
-        bindir = self._stub(tmp_path, {primary: _GOOD_REPORT, secondary: _GOOD_REPORT})
-        outputs = _run_action(tmp_path, self._env(tmp_path, primary, secondary), bindir)
-        assert outputs.get("verdict") == "COMPATIBLE", outputs
-        assert outputs["_exit"] == 0, outputs
-
-    def test_a_stale_destination_is_not_read_as_this_runs_verdict(
-        self, tmp_path: Path
-    ) -> None:
-        # The forgery the freshness check exists to stop, stated as its own
-        # case: a PR-committed report claiming COMPATIBLE must not be able to
-        # answer for a run that wrote nothing at all.
-        primary, secondary = self._paths(tmp_path)
-        primary.write_text(_GOOD_REPORT, encoding="utf-8")
-        secondary.write_text(_GOOD_REPORT, encoding="utf-8")
-        bindir = self._stub(tmp_path, {})
-        outputs = _run_action(tmp_path, self._env(tmp_path, primary, secondary), bindir)
-        assert outputs.get("verdict") == "REPORT_UNREADABLE", outputs
-        assert outputs["_exit"] == 1, outputs
-
-
-#: Workflow-command payloads, embedded in a *destination path* rather than in
-#: report content. `extra-args` is PR-controlled per `action/AGENTS.md`'s threat
-#: model, so `--write json=<payload>` is a caller-supplied string that reaches a
-#: `::error::` annotation whenever that destination fails to arrive.
-#:
-#: `%0A` is the one that matters and the one that was missed: GitHub
-#: percent-decodes workflow-command *data*, so a raw `%0A` in the emitted line
-#: becomes a newline on the runner and everything after it is parsed as a
-#: second command. A literal newline cannot survive a shell path here; a
-#: percent-encoded one can.
-DESTINATION_INJECTION_PAYLOADS = (
-    "x%0A::add-mask::secret",
-    "x%0A::set-output name=verdict::COMPATIBLE",
-    "x%0A::error::forged",
-    "x%0A::stop-commands::token",
-    "x%0D::notice::forged",
-    "x%25%30A::error::double-encoded",
-)
-
-
-class TestADestinationPathCannotForgeAWorkflowCommand:
-    """The path in the diagnostic is attacker-controlled, so attack it.
-
-    Written as an executed attack rather than an assertion over `run.sh`'s
-    text, for the reason this repository already paid for once (#705 -> #758):
-    a defense asserted by reading the defending file passed while the next
-    payload went straight through it.
-
-    The invariant: for any caller-supplied destination path, no line this step
-    emits may contain a decodable workflow-command separator that the path put
-    there. `_sanitize_annotation` escapes `%` first and then flattens CR/LF, so
-    a `%0A` in the input reaches the log as the literal text `%250A` --
-    displayed, never decoded.
-    """
-
-    @pytest.mark.parametrize("payload", DESTINATION_INJECTION_PAYLOADS)
-    def test_no_raw_percent_escape_from_the_path_reaches_the_log(
-        self, tmp_path: Path, payload: str
-    ) -> None:
-        dest = tmp_path / payload
-        bindir = self._stub_that_writes_nothing(tmp_path)
-        outputs = _run_action(tmp_path, self._env(tmp_path, dest), bindir)
-        # The step must still fail -- the attack must not be prevented by the
-        # destination quietly being treated as satisfied.
-        assert outputs.get("verdict") == "REPORT_UNREADABLE", outputs
-        errors = [
-            line
-            for line in outputs["_stdout"].splitlines()
-            if line.startswith("::error::")
-        ]
-        assert errors, outputs["_stdout"]
-        for line in errors:
-            assert "%0A" not in line.upper().replace("%0D", "%0A") or "%25" in line, (
-                payload,
-                line,
-            )
-            # The decisive check: the payload's own escape must appear escaped.
-            assert payload not in line, (payload, line)
-            assert "%250A" in line or "%250D" in line or "%25" in line, (payload, line)
-
-    @pytest.mark.parametrize("payload", DESTINATION_INJECTION_PAYLOADS)
-    def test_the_forged_command_never_becomes_its_own_line(
-        self, tmp_path: Path, payload: str
-    ) -> None:
-        # Even if a future change decoded before emitting, the smuggled command
-        # must not end up as a line of its own that a runner would execute.
-        dest = tmp_path / payload
-        bindir = self._stub_that_writes_nothing(tmp_path)
-        outputs = _run_action(tmp_path, self._env(tmp_path, dest), bindir)
-        smuggled = payload.split("%0A")[-1].split("%0D")[-1]
-        for line in outputs["_stdout"].splitlines():
-            assert line.strip() != smuggled.strip(), (payload, line)
-
-    def test_an_ordinary_path_is_still_named_in_full(self, tmp_path: Path) -> None:
-        # Narrowness control: sanitizing must not mangle the common case, or
-        # the diagnostic stops telling the user which artifact went missing.
-        dest = tmp_path / "reports" / "abi.json"
-        dest.parent.mkdir(parents=True, exist_ok=True)
-        bindir = self._stub_that_writes_nothing(tmp_path)
-        outputs = _run_action(tmp_path, self._env(tmp_path, dest), bindir)
-        errors = [
-            line
-            for line in outputs["_stdout"].splitlines()
-            if line.startswith("::error::")
-        ]
-        assert any(str(dest) in line for line in errors), errors
-
-    def _stub_that_writes_nothing(self, tmp_path: Path) -> Path:
-        bindir = tmp_path / "bin"
-        bindir.mkdir(exist_ok=True)
-        stub = bindir / "abicheck"
-        stub.write_text("#!/usr/bin/env bash\nexit 0\n", encoding="utf-8")
-        stub.chmod(0o755)
-        return bindir
-
-    def _env(self, tmp_path: Path, dest: Path) -> dict[str, str]:
-        return {
-            "INPUT_MODE": "compare",
-            "INPUT_OLD_LIBRARY": _lib(tmp_path, "libold.so"),
-            "INPUT_NEW_LIBRARY": _lib(tmp_path, "libnew.so"),
-            "INPUT_FORMAT": "markdown",
-            "INPUT_EXTRA_ARGS": f"--write json={dest}",
-        }
-
-
-#: Strings that parse, are non-empty, and are not verdicts any consumer can act
-#: on. The reported instance was `"write interrupted"`; the class is "anything
-#: outside the emitters' own vocabulary".
-UNKNOWN_VERDICT_STRINGS = (
-    "write interrupted",
-    "compatible",
-    "OK",
-    "PASS",
-    "BREAKING_CHANGE",
-    "COMPATIBLE ",
-    "null",
-    "0",
-)
-
-
-class TestAnUnknownVerdictStringIsNotAResult:
-    """`{"verdict": "<anything>"}` must not license the COMPATIBLE fallthrough.
-
-    `_carries_a_result` accepted any non-empty verdict string, so a document
-    like `{"verdict": "write interrupted"}` read as `ok`; `compat_verdict` then
-    returned that unknown value, and `_resolve_clean_exit_verdict` -- which
-    recognizes only the break and risk tiers -- kept its initial COMPATIBLE
-    (Codex review, P2). Membership in the emitters' own vocabulary is what the
-    reader now requires.
-    """
-
-    @pytest.mark.parametrize("verdict", UNKNOWN_VERDICT_STRINGS)
-    def test_it_never_publishes_compatible(self, tmp_path: Path, verdict: str) -> None:
-        payload = json.dumps({"report_schema_version": "4.4", "verdict": verdict})
-        bindir = _stub_abicheck(tmp_path, exit_code=0, payload=payload.encode("utf-8"))
-        outputs = _run_action(tmp_path, _compare_env(tmp_path), bindir)
-        assert outputs.get("verdict") == "REPORT_UNREADABLE", (verdict, outputs)
-        assert outputs["_exit"] == 1, (verdict, outputs)
-
-    @pytest.mark.parametrize(
-        "verdict",
-        ("NO_CHANGE", "COMPATIBLE", "COMPATIBLE_WITH_RISK", "API_BREAK", "BREAKING"),
-    )
-    def test_every_real_verdict_still_reads(self, tmp_path: Path, verdict: str) -> None:
-        # The control that keeps the vocabulary from being satisfiable by
-        # rejecting everything: each tier the CLI actually emits must pass.
-        payload = json.dumps({"report_schema_version": "4.4", "verdict": verdict})
-        bindir = _stub_abicheck(tmp_path, exit_code=0, payload=payload.encode("utf-8"))
-        outputs = _run_action(tmp_path, _compare_env(tmp_path), bindir)
-        assert outputs.get("verdict") != "REPORT_UNREADABLE", (verdict, outputs)
